Clear Commnets Display
